Fix in of Zookeeper config management

+++ Site/ASAB Maestro/Files/zookeeper/confro/zkstart.sh
@@ -3,7 +3,11 @@
 # Copy the configuration from the read-only site directory to the /conf
 # This has to be done b/c Zookeeper cannot operate from read-only /conf
 # Zookeeper add own files into /conf during runtime so `/conf` cannot be mounted directly (rsync will erase Zookeeper data etc.)
-cp /confro/zoo.cfg $ZOO_CONF_DIR/zoo.cfg
+# WARNING: The file will be copied only once, in the beginning b/c Zookeeper manages the content of that file
+# If any modification will be needed, the custom "patching" functionality must me added here.
+if [[ ! -f "$ZOO_CONF_DIR/zoo.cfg" ]]; then
+    cp /confro/zoo.cfg $ZOO_CONF_DIR/zoo.cfg
+fi
 
 # Create a "myid" file in the data directory
 if [[ ! -f "$ZOO_DATA_DIR/myid" ]]; then
